requirements.txt
# Core dependencies (actualizadas para seguridad)
beautifulsoup4==4.12.3
lxml==5.3.0
pyTelegramBotAPI==4.21.0
python-dotenv==1.0.1
requests==2.32.3

# Security
cryptography==42.0.5